[get-reststops-by-highways] Menu refactoring
