+++ core/src/main/kotlin/com/eightpotatoes/nexters/core/model/ReststopFood.kt
@@ -36,7 +36,7 @@ data class ReststopFood(
             isRecommended = recommendyn.equals("Y", ignoreCase = true),
             isPremium = premiumyn.equals("Y", ignoreCase = true),
             isBestFood = bestfoodyn.equals("Y", ignoreCase = true),
-            reststopId = stdRestCd
+            reststopCode = stdRestCd
         )
     }
 }
