+++ core/src/main/kotlin/com/eightpotatoes/nexters/core/repository/MenuRepository.kt
@@ -4,5 +4,5 @@ import com.eightpotatoes.nexters.core.entity.Menu
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface MenuRepository : JpaRepository<Menu, Int>, MenuRepositoryCustom {
-    fun findByReststopId(reststopId: String?): List<Menu>
+    fun findByReststopCode(reststopCode: String?): List<Menu>
 }
