+++ core/src/main/kotlin/com/eightpotatoes/nexters/core/repository/MenuRepositoryCustomImpl.kt
@@ -4,6 +4,7 @@ import com.eightpotatoes.nexters.core.entity.Menu
 import jakarta.persistence.EntityManager
 import jakarta.persistence.PersistenceContext
 import jakarta.transaction.Transactional
+import java.time.LocalDateTime
 
 open class MenuRepositoryCustomImpl : MenuRepositoryCustom {
 
@@ -12,30 +13,35 @@ open class MenuRepositoryCustomImpl : MenuRepositoryCustom {
 
     @Transactional
     override fun upsert(menu: Menu) {
+        if (menu.createdAt == null) {
+            menu.createdAt = LocalDateTime.now()
+        }
+        menu.modifiedAt = LocalDateTime.now()
+
         val sql = """
-            INSERT INTO menu (description, food_seq, is_best_food, is_premium, is_recommended, name, price, reststop_id, created_at, modified_at) 
-            VALUES (:description, :foodSeq, :isBestFood, :isPremium, :isRecommended, :name, :price, :reststopId, :createdAt, :modifiedAt) 
+            INSERT INTO menu (description, food_seq, is_best_food, is_premium, is_recommended, name, price, reststop_code, created_at, modified_at) 
+            VALUES (:description, :foodSeq, :isBestFood, :isPremium, :isRecommended, :name, :price, :reststopCode, :createdAt, :modifiedAt) 
             ON DUPLICATE KEY UPDATE 
             description = VALUES(description), 
             is_best_food = VALUES(is_best_food), 
             is_premium = VALUES(is_premium), 
             is_recommended = VALUES(is_recommended), 
-            modified_at = VALUES(modified_at), 
             price = VALUES(price), 
-            food_seq = VALUES(food_seq)
+            food_seq = VALUES(food_seq),
+            modified_at = VALUES(modified_at)
         """
 
         val query = entityManager.createNativeQuery(sql).apply {
-            setParameter("createdAt", menu.createdAt)
+            setParameter("name", menu.name)
+            setParameter("reststopCode", menu.reststopCode)
             setParameter("description", menu.description)
-            setParameter("foodSeq", menu.foodSeq)
             setParameter("isBestFood", menu.isBestFood)
             setParameter("isPremium", menu.isPremium)
             setParameter("isRecommended", menu.isRecommended)
-            setParameter("modifiedAt", menu.modifiedAt)
-            setParameter("name", menu.name)
             setParameter("price", menu.price)
-            setParameter("reststopId", menu.reststopId)
+            setParameter("foodSeq", menu.foodSeq)
+            setParameter("createdAt", menu.createdAt)
+            setParameter("modifiedAt", menu.modifiedAt)
         }
         query.executeUpdate()
     }
